requirements.txt
request
javalang
ollama
pandas
